// packages/datadog-api-client-v1/models/SyntheticsAPIStep.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SyntheticsAPIStepSubtype } from './SyntheticsAPIStepSubtype';
import { SyntheticsAssertion } from './SyntheticsAssertion';
import { SyntheticsParsingOptions } from './SyntheticsParsingOptions';
import { SyntheticsTestRequest } from './SyntheticsTestRequest';
/**
* The steps used in a Synthetics multistep API test.
*/
export declare class SyntheticsAPIStep {
    /**
    * Determines whether or not to continue with test if this step fails.
    */
    'allowFailure'?: boolean;
    /**
    * Array of assertions used for the test.
    */
    'assertions'?: Array<SyntheticsAssertion>;
    /**
    * Array of values to parse and save as variables from the response.
    */
    'extractedValues'?: Array<SyntheticsParsingOptions>;
    /**
    * Determines whether or not to consider the entire test as failed if this step fails. Can be used only if `allowFailure` is `true`.
    */
    'isCritical'?: boolean;
    /**
    * The name of the step.
    */
    'name'?: string;
    'request'?: SyntheticsTestRequest;
    'subtype'?: SyntheticsAPIStepSubtype;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsAPIStep;
    static serialize(data: SyntheticsAPIStep): {
        [key: string]: any;
    };
    constructor();
}