// packages/datadog-api-client-v1/models/SyntheticsConfigVariable.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SyntheticsConfigVariableType } from './SyntheticsConfigVariableType';
/**
* Object defining a variable that can be used in your test configuration.
*/
export declare class SyntheticsConfigVariable {
    /**
    * Example for the variable.
    */
    'example': string;
    /**
    * Name of the variable.
    */
    'name': string;
    /**
    * Pattern of the variable.
    */
    'pattern'?: string;
    'type': SyntheticsConfigVariableType;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsConfigVariable;
    static serialize(data: SyntheticsConfigVariable): {
        [key: string]: any;
    };
    constructor();
}